Changing function types

+++ TP6/TP6_Project/Project/pid_regulator.h
@@ -5,6 +5,6 @@
 //start the PID regulator thread
 void pid_regulator_start(void);
 
-void motor_set_position(float position_r, float position_l, float speed_r, float speed_l);
+void motor_set_position(int16_t position_r, int16_t position_l, int16_t speed_r, int16_t speed_l);
 
 #endif /* PID_REGULATOR_H_ */
